distributed/requirements.txt
adafruit-circuitpython-dht==3.7.8
RPi.GPIO==0.7.0